Basic Main Menu
I added in main menu with buttons to play game, go to options menu, and quit game. To test this functionality, hit the play button in Unity on the top and once the game loads click on the buttons. Play should go to the basic game scene we have set up (it literally looks like dogshit right now, just like a couple green tiles). Options should go to an option menu with a slider for volume and a back button, volume slider does nothing besides slide right now. Quit should quit the play session and give a debug quit message in the console.

I also imported in an extension called textmeshpro, this is the reason for the mass of files included in this change.