Deselect buttons, next day button, general menu redesign
there's now buttons to deselect breeding selections and to move on to the next day. I repositioned the trait lists, made them bigger, and defaulted the scroll wheels to the top of the list. fixed the drones scroll wheel as well. All that is left to do is add food UI and maybe make the trait text look a bit better.