Proper menu wipes for re-opening and drone children interactions
Children are now added back to the list but become un-interactable in the breeding menu.

The final breeding button becomes uninteractable when there isn't enough food (food UI and functionality only partially added).

The drone list now properly wipes on close and the menu resets when it is opened more than once.