Combat start and drone tile traversal movement
Added a movement queue for drones, an animation to walk drones to the next tile (with automatic standing spots generated by the number of drones) and spawned enemies to walk in from the opposite door of the room. Disabled collision of ally drones with each other. Breeding animation is broken for this commit.